requirements.txt
requests
Pillow
moviepy
python-dotenv
google-generativeai
diffusers
torch
selenium
pyautogui
pytz
toml
webdriver-manager